start moving (obsolete) quicksurvey code to dedicated module
